+++ stories/setup/styles.scss
@@ -2,11 +2,11 @@ body {
   padding: 24px;
 }
 
-.CodeShapeStory {
+.CodeLineShapesStory {
   display: flex;
 }
 
-.CodeShapeStory-panel {
+.CodeLineShapesStory-panel {
   border: 1px solid;
   padding: 10px;
 
@@ -15,13 +15,13 @@ body {
   }
 }
 
-.CodeShapeStory-lineShape {
+.CodeLineShapesStory-lineShape {
   display: flex;
   width: 100%;
 }
 
-.CodeShapeStory-actualCode,
-.CodeShapeStory-lineShape {
+.CodeLineShapesStory-actualCode,
+.CodeLineShapesStory-lineShape {
   height: 1rem;
   line-height: 1;
   margin-top: 2px;
@@ -31,7 +31,7 @@ body {
   }
 }
 
-.CodeShapeStory-actualCode {
+.CodeLineShapesStory-actualCode {
   overflow: hidden;
 
   pre {
